added OSLog+Utils
