+++ PokeApi/Networking/ServiceManager.swift
@@ -82,13 +82,3 @@ struct ServiceManger {
         }
     }
 }
-
-
-extension OSLog {
-    
-    private static var subsystem = Bundle.main.bundleIdentifier!
-    static let navigation = OSLog(subsystem: subsystem, category: "navigation")
-    static let businessLogic = OSLog(subsystem: subsystem, category: "buisnessLogic")
-    static let view = OSLog(subsystem: subsystem, category: "view")
-    static let network = OSLog(subsystem: subsystem, category: "networking")
-}
